Refactor enhancements/__init__.py to submodules
